refactor: (#109) 로그인 응답에서 유저 프로필 조회 API 분리
