+++ src/auth/dtos/responses/login-response.dto.ts
@@ -1,10 +1,3 @@
-import { OutfitAssetPaths } from 'src/outfit/outfit.type';
-
 export class LoginResponseDto {
   accessToken: string;
-  tag: string;
-  coins: number;
-  level: number;
-  currentExp: number;
-  outfit: OutfitAssetPaths;
 }
