+++ src/user/dtos/responses/profile-response.dto.ts
@@ -1,0 +1,9 @@
+import { OutfitAssetPaths } from 'src/outfit/outfit.type';
+
+export class ProfileResponseDto {
+  tag: string;
+  coins: number;
+  level: number;
+  currentExp: number;
+  outfit: OutfitAssetPaths;
+}
