+++ src/user/user.module.ts
@@ -3,9 +3,10 @@ import { UserService } from './user.service';
 import { UserRepository } from './user.repository';
 import { UserController } from './user.controller';
 import { PrismaModule } from 'src/prisma/prisma.module';
+import { OutfitModule } from 'src/outfit/outfit.module';
 
 @Module({
-  imports: [PrismaModule],
+  imports: [PrismaModule, OutfitModule],
   controllers: [UserController],
   providers: [UserService, UserRepository],
   exports: [UserService],
